+++ Examples/Benchmark/Package.swift
@@ -3,7 +3,7 @@
 import PackageDescription
 
 let package = Package(
-    name: "MyLambda",
+    name: "swift-aws-lambda-runtime-example",
     platforms: [
         .macOS(.v12),
     ],
@@ -14,15 +14,13 @@ let package = Package(
         // this is the dependency on the swift-aws-lambda-runtime library
         // in real-world projects this would say
         // .package(url: "https://github.com/swift-server/swift-aws-lambda-runtime.git", from: "1.0.0")
-        .package(name: "swift-aws-lambda-runtime", path: "../../.."),
-        .package(name: "Shared", path: "../Shared"),
+        .package(name: "swift-aws-lambda-runtime", path: "../.."),
     ],
     targets: [
         .executableTarget(
             name: "MyLambda",
             dependencies: [
-                .product(name: "AWSLambdaRuntime", package: "swift-aws-lambda-runtime"),
-                .product(name: "Shared", package: "Shared"),
+                .product(name: "AWSLambdaRuntimeCore", package: "swift-aws-lambda-runtime"),
             ],
             path: "."
         ),
