+++ Examples/Deployment/Dockerfile
@@ -1,0 +1,3 @@
+FROM swift:5.5-amazonlinux2
+
+RUN yum -y install zip
