+++ Examples/Deployment/scripts/build-and-package.sh
@@ -1,0 +1,39 @@
+#!/bin/bash
+##===----------------------------------------------------------------------===##
+##
+## This source file is part of the SwiftAWSLambdaRuntime open source project
+##
+## Copyright (c) 2020 Apple Inc. and the SwiftAWSLambdaRuntime project authors
+## Licensed under Apache License v2.0
+##
+## See LICENSE.txt for license information
+## See CONTRIBUTORS.txt for the list of SwiftAWSLambdaRuntime project authors
+##
+## SPDX-License-Identifier: Apache-2.0
+##
+##===----------------------------------------------------------------------===##
+
+set -eu
+
+executable=$1
+workspace="$(pwd)/../.."
+
+echo "-------------------------------------------------------------------------"
+echo "preparing docker build image"
+echo "-------------------------------------------------------------------------"
+docker build . -t builder
+echo "done"
+
+echo "-------------------------------------------------------------------------"
+echo "building \"$executable\" lambda"
+echo "-------------------------------------------------------------------------"
+docker run --rm -v "$workspace":/workspace -w /workspace/Examples/Deployment builder \
+       bash -cl "swift build --product $executable -c release"
+echo "done"
+
+echo "-------------------------------------------------------------------------"
+echo "packaging \"$executable\" lambda"
+echo "-------------------------------------------------------------------------"
+docker run --rm -v "$workspace":/workspace -w /workspace/Examples/Deployment builder \
+       bash -cl "./scripts/package.sh $executable"
+echo "done"
