+++ Examples/Deployment/scripts/serverless-deploy.sh
