+++ Examples/Deployment/scripts/serverless-remove.sh
