+++ Examples/JSON/Lambda.swift
@@ -1,0 +1,41 @@
+//===----------------------------------------------------------------------===//
+//
+// This source file is part of the SwiftAWSLambdaRuntime open source project
+//
+// Copyright (c) 2021 Apple Inc. and the SwiftAWSLambdaRuntime project authors
+// Licensed under Apache License v2.0
+//
+// See LICENSE.txt for license information
+// See CONTRIBUTORS.txt for the list of SwiftAWSLambdaRuntime project authors
+//
+// SPDX-License-Identifier: Apache-2.0
+//
+//===----------------------------------------------------------------------===//
+
+import AWSLambdaRuntime
+
+struct Request: Codable {
+    let body: String
+}
+
+struct Response: Codable {
+    let body: String
+}
+
+// in this example we are receiving and responding with codables. Request and Response above are examples of how to use
+// codables to model your request and response objects
+
+@main
+struct MyLambda: LambdaHandler {
+    typealias Event = Request
+    typealias Output = Response
+
+    init(context: Lambda.InitializationContext) async throws {
+        // setup your resources that you want to reuse for every invocation here.
+    }
+
+    func handle(_ event: Request, context: Lambda.Context) async throws -> Response {
+        // as an example, respond with the input event's reversed body
+        Response(body: String(event.body.reversed()))
+    }
+}
