--- Examples/LambdaFunctions/Dockerfile
@@ -1,3 +1,0 @@
-FROM swift:5.2-amazonlinux2
-
-RUN yum -y install zip
