--- Examples/LambdaFunctions/scripts/serverless/CurrencyExchange-template.yml
@@ -1,20 +1,0 @@
-service: currency-swift-aws
-
-package:
-    artifact: .build/lambda/CurrencyExchange/lambda.zip
-
-provider:
-    name: aws
-    runtime: provided
-    iamRoleStatements:
-        - Effect: Allow
-          Action:
-              - logs:CreateLogGroup
-              - logs:CreateLogStream
-              - logs:PutLogEvents
-          Resource: "*"        
-
-functions:
-    currencyExchangeFunction:
-        handler: CurrencyExchange
-        memorySize: 128
