--- Examples/LambdaFunctions/scripts/serverless/ErrorHandling-template.yml
@@ -1,20 +1,0 @@
-service: errorhandling-swift-aws
-
-package:
-    artifact: .build/lambda/ErrorHandling/lambda.zip
-
-provider:
-    name: aws
-    runtime: provided
-    iamRoleStatements:
-        - Effect: Allow
-          Action:
-              - logs:CreateLogGroup
-              - logs:CreateLogStream
-              - logs:PutLogEvents
-          Resource: "*"        
-
-functions:
-    errorHandlingFunction:
-        handler: ErrorHandling
-        memorySize: 128
