--- Examples/LocalDebugging/MyLambda/Tests/LinuxMain.swift
@@ -1,15 +1,0 @@
-//===----------------------------------------------------------------------===//
-//
-// This source file is part of the SwiftAWSLambdaRuntime open source project
-//
-// Copyright (c) 2020 Apple Inc. and the SwiftAWSLambdaRuntime project authors
-// Licensed under Apache License v2.0
-//
-// See LICENSE.txt for license information
-// See CONTRIBUTORS.txt for the list of SwiftAWSLambdaRuntime project authors
-//
-// SPDX-License-Identifier: Apache-2.0
-//
-//===----------------------------------------------------------------------===//
-
-preconditionFailure("use `swift test --enable-test-discovery`")
