+++ docker/docker-compose.yaml
@@ -1,6 +1,6 @@
 # this file is not designed to be run directly
 # instead, use the docker-compose.<os>.<swift> files
-# eg docker-compose -f docker/docker-compose.yaml -f docker/docker-compose.al2.52.yaml run test
+# eg docker-compose -f docker/docker-compose.yaml -f docker/docker-compose.al2.55.yaml run test
 version: "3"
 
 services:
@@ -34,8 +34,15 @@ services:
     <<: *common
     command: >-
       /bin/bash -clx "
-      swift build --package-path Examples/LambdaFunctions &&
-      swift build --package-path Examples/LocalDebugging/MyLambda"
+      swift build --package-path Examples/Benchmark &&
+      swift build --package-path Examples/Deployment &&
+      swift build --package-path Examples/Echo &&
+      swift build --package-path Examples/ErrorHandling &&
+      swift build --package-path Examples/Foundation &&
+      swift build --package-path Examples/JSON &&
+      swift build --package-path Examples/LocalDebugging/MyLambda  &&
+      swift test --package-path Examples/Testing
+      "
 
   # util
